Add prim.svd, refs.linalg.svd, and refs.linalg.svdvals
This is the first prim / ref added that has multiple returns.
There is an issue with `out_wrapper_multi` as currently implemented
(left a note). It assumes that the API is `svd(X, U=U, S=S, Vh=Vh)`,
when it's actually `svd(X, out=(U, S, Vh))`.

Even more, if we want to model PyTorch exactly, it should return a
`torch.return_types.svd`, rather than a `Tuple`.

There is an issue with
As per title

ghstack-source-id: e2e00282924ea3d728a9ab2e8e6c2d7618bc0f9a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/78616